requirements.txt
python-dotenv
pydrive2
appdirs
fsspec
pandas
funcy
tqdm
joblib
